+++ test/PanelTests.cs
@@ -17,7 +17,7 @@ namespace Elements.Tests
             var c = new Vector3(1,0,1);
             var d = new Vector3(0,0,1);
             var panel = new Panel(new []{a,b,c,d}, BuiltInMaterials.Glass);
-            Assert.Equal(BuiltInMaterials.Glass, panel.Geometry[0].Material);
+            Assert.Equal(BuiltInMaterials.Glass, panel.Material);
             // Assert.Equal(panel.Geometry[0].Faces[0].Vertices, panel.Perimeter);
             this.Model.AddElement(panel);
         }